switch project

+++ src/index.js
@@ -18,32 +18,29 @@
 import {ToDo} from "./newToDo.js";
 import {markComplete} from "./markComplete.js";
 import {deleteItem} from "./deleteItem.js";
-import {createProject} from "./createProject.js";
-import {Project} from "./newProject.js";
 
 
 const form  = document.getElementById('addform');
 const projectForm  = document.getElementById('addprojectform');
 
 var toDoList = [{title: '01hello', dueDate: '2022-07-21', completed: false, assignedProject: "Coding"}, {title: '02hello', dueDate: '2022-07-21', completed: true, assignedProject: "Inbox"}, {title: '03hello', dueDate: '2022-07-21', completed: false, assignedProject: "Coding"}];
-var projects = [{projectName: "Due Today"}, {projectName: "Coding"}]
+var projects = ["Inbox", "Due Today", "Coding"]
+
+let chosenProject = projects[0]
 
 projectForm.addEventListener('submit', (event) => {
   event.preventDefault();
-  console.log("hwllo")
   let projectName = projectForm.elements["project"].value
-  var Project1 = new Project(projectName)
-  projects.push(Project1);
-  console.log(projects)
+  projects.push(projectName);
   displayProjects();
 });
 
-
-
-displayList();
+displayList(chosenProject);
 displayProjects();
 
-function displayList(){
+function displayList(chosenProject){
+
+  console.log(chosenProject)
 
   form.addEventListener('submit', (event) => {
     event.preventDefault();
@@ -53,7 +50,7 @@ function displayList(){
     var ToDo1 = new ToDo(title, dueDate, false, assignedProject)
     toDoList.push(ToDo1);
     console.log(toDoList)
-    displayList();
+    displayList(chosenProject);
   });
 
   const listContainer = document.getElementById("todos");
@@ -63,61 +60,65 @@ function displayList(){
   }
 
   for (let i = 0; i < toDoList.length; i++){
-    let itemContainer = document.createElement("div");
-    itemContainer.className = "todo";
-    
-    let item = document.createElement("div");
-    item.className = "todoname";
-    item.innerHTML = toDoList[i].title;
-    
-    let completeStatus = toDoList[i].completed;
-    item.addEventListener("click", function() { markComplete(completeStatus, i, toDoList); } , false)
 
-    if(toDoList[i].completed === true){
-      item.classList = "completed";
+    if (toDoList[i].assignedProject == chosenProject){
+      let itemContainer = document.createElement("div");
+      itemContainer.className = "todo";
+      
+      let item = document.createElement("div");
+      item.className = "todoname";
+      item.innerHTML = toDoList[i].title;
+      
+      let completeStatus = toDoList[i].completed;
+      item.addEventListener("click", function() { markComplete(completeStatus, i, toDoList); } , false)
+
+      if(toDoList[i].completed === true){
+        item.classList = "completed";
+      }
+
+      let itemDate = document.createElement("div");
+      itemDate.className = "dueDate";
+      itemDate.innerHTML = toDoList[i].dueDate;
+
+      let itemDelete = document.createElement("img");
+      itemDelete.src = "img/trash-can.png";
+      itemDelete.className = "itemDelete";
+      itemDelete.addEventListener("click", function() { deleteItem(i, toDoList); } , false)
+      itemDelete.width = 25;
+
+      let itemLeft = document.createElement("div");
+      itemLeft.className = "itemLeft";
+
+      itemContainer.appendChild(item);
+      itemLeft.appendChild(itemDate);
+      itemLeft.appendChild(itemDelete);
+      itemContainer.appendChild(itemLeft);
+      listContainer.appendChild(itemContainer);
     }
-
-    let itemDate = document.createElement("div");
-    itemDate.className = "dueDate";
-    itemDate.innerHTML = toDoList[i].dueDate;
-
-    let itemDelete = document.createElement("img");
-    itemDelete.src = "img/trash-can.png";
-    itemDelete.className = "itemDelete";
-    itemDelete.addEventListener("click", function() { deleteItem(i, toDoList); } , false)
-    itemDelete.width = 25;
-
-    let itemLeft = document.createElement("div");
-    itemLeft.className = "itemLeft";
-
-    itemContainer.appendChild(item);
-    itemLeft.appendChild(itemDate);
-    itemLeft.appendChild(itemDelete);
-    itemContainer.appendChild(itemLeft);
-    listContainer.appendChild(itemContainer);
   }
     
 }
 
 function displayProjects(){
-  
-
   const projectContainer = document.getElementById("projectlist");
   while (projectContainer.firstChild) {
     projectContainer.removeChild(projectContainer.lastChild);
   }
+
   for (let i = 0; i < projects.length; i++){
     let projectName = document.createElement("h3");
-    projectName.innerHTML = projects[i].projectName;
-    let nametest = projects[i].projectName
-    projectName.addEventListener("click", function() { switchProject(nametest); } , false)
+    projectName.innerHTML = projects[i];
+    let projectID = projects[i];
+    projectName.addEventListener("click", function() { switchProject(projectID); } , false)
     projectContainer.appendChild(projectName);
   }
 
 }
 
-function switchProject(nametest){
-  console.log(nametest)
+function switchProject(projectID){
+  chosenProject = projectID
+  console.log(chosenProject)
+  displayList(chosenProject)
 }
 
 export {displayList}
